docs: Refactors closures to tests; Adds function pointers and more on closures
